Move factory and fix input type
